completed and tested all routes

+++ backend/routes/api/bookings.js
@@ -3,7 +3,7 @@ const router = require('express').Router();
 const { Op } = require('sequelize')
 const { check } = require('express-validator');
 const { handleValidationErrors } = require('../../utils/validation');
-const { Booking, Spot } = require('../../db/models');
+const { Booking, Spot, SpotImage } = require('../../db/models');
 const { authCheck } = require('../../utils/auth');
 
 //  REQ AUTH - Get All Current User's Bookings
@@ -14,14 +14,28 @@ router.get('/current', authCheck, async (req, res, next) => {
         where: {
             userId: user.id
         },
-        include: { model: Spot }
+        include: {
+            model: Spot,
+            attributes: ['id', 'ownerId', 'address', 'city', 'state', 'country', 'lat', 'lng', 'name', 'price'],
+            include: {
+                model: SpotImage,
+                where: {
+                    preview: true
+                }
+            }
+        }
     })
 
-    let Bookings = JSON.stringify(bookings)
-    Bookings = JSON.parse(Booking)
+    bookings = JSON.stringify(bookings)
+    bookings = JSON.parse(bookings)
 
-    return res.json({ Bookings })
-    // res.json({ route: "get/bookings/current" })
+    for (let i = 0; i < bookings.length; i++) {
+        bookings[i].Spot.previewImage = bookings[i].Spot.SpotImages[0].url;
+        delete bookings[i].Spot.SpotImages
+    }
+
+    return res.json({ Bookings: bookings })
+
 })
 
 // REQ AUTH - Edit a Booking
@@ -29,9 +43,6 @@ router.put('/:bookingId', authCheck, async (req, res, next) => {
     const { user } = req;
     const { bookingId } = req.params;
     const { startDate, endDate } = req.body;
-
-
-
 
     const booking = await Booking.findOne({
         where: {
@@ -56,12 +67,8 @@ router.put('/:bookingId', authCheck, async (req, res, next) => {
     }
 
     if (bookingEdit.startDate >= bookingEdit.endDate) {
-        // create err 
         const err = new Error("Bad Request")
-        // set error title
-        // set status code
         err.status = 400;
-        // next(err)
         err.errors = {};
         err.errors.endDate = "endDate cannot be on or before startDate";
         return next(err);
@@ -105,65 +112,39 @@ router.put('/:bookingId', authCheck, async (req, res, next) => {
     })
 
     if (conflicts.length > 0) {
-        // do error thingo
         const err = new Error('Sorry, this spot is already booked for the specified dates');
         err.errors = {};
         err.status = 403;
 
         for (let i = 0; i < conflicts.length; i++) {
-            // console.log('start date', conflicts[0].Bookings[0].startDate)
-            // console.log('end date', conflicts[0].Bookings[0].endDate)
-            // console.log('bookingEntry start date', new Date(bookingEntry.startDate))
-            // console.log('start date test', new Date(bookingEntry.startDate) >= conflicts[0].Bookings[0].startDate && new Date(bookingEntry.startDate) <= conflicts[0].Bookings[0].endDate)
             if (new Date(bookingEdit.startDate) >= conflicts[i].startDate && new Date(bookingEdit.startDate) <= conflicts[i].endDate) {
-                // console.log('start date err')
-
-                // next(err)
-
                 err.errors.startDate = "Start date conflicts with an existing booking";
 
-            }              // console.log('start date test', bookingEntry.startDate == conflicts[0].Bookings[0].startDate)
+            }
         }
 
         for (let i = 0; i < conflicts.length; i++) {
-            // console.log('start date', conflicts[0].Bookings[0].startDate)
-            // console.log('end date', conflicts[0].Bookings[0].endDate)
-            // console.log('bookingEntry start date', new Date(bookingEntry.startDate))
-            // console.log('start date test', new Date(bookingEntry.startDate) >= conflicts[0].Bookings[0].startDate && new Date(bookingEntry.startDate) <= conflicts[0].Bookings[0].endDate)
             if (new Date(bookingEdit.endDate) >= conflicts[i].startDate && new Date(bookingEdit.endDate) <= conflicts[i].endDate) {
-                // console.log('end date err')
                 err.errors.endDate = "End date conflicts with an existing booking";
-            }              // console.log('start date test', bookingEntry.startDate == conflicts[0].Bookings[0].startDate)
+            }
         }
 
         for (let i = 0; i < conflicts.length; i++) {
-            // console.log('start date', conflicts[0].Bookings[0].startDate)
-            // console.log('end date', conflicts[0].Bookings[0].endDate)
-            // console.log('bookingEntry start date', new Date(bookingEntry.startDate))
-            // console.log('start date test', new Date(bookingEntry.startDate) >= conflicts[0].Bookings[0].startDate && new Date(bookingEntry.startDate) <= conflicts[0].Bookings[0].endDate)
             if (
                 new Date(bookingEdit.endDate) > conflicts[i].startDate
                 && new Date(bookingEdit.endDate) > conflicts[i].endDate
                 && new Date(bookingEdit.startDate) < conflicts[i].startDate
                 && new Date(bookingEdit.startDate) < conflicts[i].endDate
             ) {
-                // console.log('booking surrounds date err')
                 err.errors.startDate = "Start date conflicts with an existing booking";
                 err.errors.endDate = "End date conflicts with an existing booking";
-            }              // console.log('start date test', bookingEntry.startDate == conflicts[0].Bookings[0].startDate)
+            }
         }
         return next(err);
     }
 
-
-
-    // console.log(conflicts, 'bookingId', bookingId)
-
     const updatedBooking = await booking.update(bookingEdit)
-
-    // return res.json(conflicts)
     return res.json(updatedBooking)
-    // res.json({ route: "put/bookings/:bookingId" })
 })
 
 // REQ AUTH - Delete a Booking
@@ -189,6 +170,12 @@ router.delete('/:bookingId', authCheck, async (req, res, next) => {
         return next(err)
     }
 
+    if (new Date(booking.startDate) < Date.now()) {
+        const err = new Error("Bookings that have been started can't be deleted")
+        err.status = 403;
+        return next(err)
+    }
+
     await Booking.destroy({
         where: {
             id: bookingId
@@ -196,7 +183,6 @@ router.delete('/:bookingId', authCheck, async (req, res, next) => {
     });
 
     return res.json({ "message": "Successfully deleted" })
-    // return res.json({ route: "delete/bookings/:bookingId" })
 })
